add jammie expedition and option to exit game after regen

+++ bh_bot/classes/image_info.py
@@ -10,5 +10,5 @@ class ImageInfo:
     offset_y: int = 0
     clicks: int = 1
     optional: bool = True
-    confidence: int = 0.8
+    confidence: float = 0.8
     grayscale: bool = True
